remove rhs tabs
