+++ webapp/src/components/rhs/rhs_about.tsx
@@ -2,7 +2,6 @@
 // See LICENSE.txt for license information.
 
 import React from 'react';
-import Scrollbars from 'react-custom-scrollbars';
 import styled from 'styled-components';
 import {useDispatch} from 'react-redux';
 
@@ -10,32 +9,12 @@ import {PlaybookRun} from 'src/types/playbook_run';
 
 import {setOwner} from 'src/client';
 import ProfileSelector from 'src/components/profile/profile_selector';
-
 import './playbook_run_details.scss';
-import {
-    renderThumbHorizontal,
-    renderThumbVertical,
-    renderView,
-} from 'src/components/rhs/rhs_shared';
 import PostCard from 'src/components/rhs/post_card';
 import {useLatestUpdate, useProfilesInCurrentChannel} from 'src/hooks';
 import PostText from 'src/components/post_text';
 import {updateStatus} from 'src/actions';
-
 import Duration from '../duration';
-
-const Description = styled.div`
-    padding: 0 0 14px 0;
-`;
-
-const Row = styled.div`
-    padding: 0 0 24px 0;
-`;
-
-const NoDescription = styled.div`
-    color: rgba(var(--center-channel-color-rgb), 0.64);
-    margin-bottom: 10px;
-`;
 
 interface Props {
     playbookRun: PlaybookRun;
@@ -73,55 +52,58 @@ const RHSAbout = (props: Props) => {
     };
 
     return (
-        <Scrollbars
-            autoHide={true}
-            autoHideTimeout={500}
-            autoHideDuration={500}
-            renderThumbHorizontal={renderThumbHorizontal}
-            renderThumbVertical={renderThumbVertical}
-            renderView={renderView}
-            style={{position: 'absolute'}}
-        >
-            <div className='PlaybookRunDetails'>
-                <Description>
-                    <div className='title'>
-                        {'Description'}
-                    </div>
-                    {description}
-                </Description>
-                <Row>
-                    <div className='side-by-side'>
-                        <div className='inner-container first-container'>
-                            <div className='first-title'>{'Owner'}</div>
-                            <ProfileSelector
-                                selectedUserId={props.playbookRun.owner_user_id}
-                                placeholder={'Assign the owner role'}
-                                placeholderButtonClass={'NoAssignee-button'}
-                                profileButtonClass={'Assigned-button'}
-                                enableEdit={true}
-                                getUsers={fetchUsers}
-                                onSelectedChange={onSelectedProfileChange}
-                                selfIsFirstOption={true}
-                            />
-                        </div>
-                        <div className='first-title'>
-                            {'Duration'}
-                            <Duration
-                                from={props.playbookRun.create_at}
-                                to={props.playbookRun.end_at}
-                            />
-                        </div>
-                    </div>
-                </Row>
-                <div id={'playbookRunRHSUpdates'}>
-                    <div className='title'>
-                        {'Recent update:'}
-                    </div>
-                    <PostCard post={latestUpdatePost}/>
+        <div className='PlaybookRunDetails'>
+            <Description>
+                <div className='title'>
+                    {'Description'}
                 </div>
+                {description}
+            </Description>
+            <Row>
+                <div className='side-by-side'>
+                    <div className='inner-container first-container'>
+                        <div className='first-title'>{'Owner'}</div>
+                        <ProfileSelector
+                            selectedUserId={props.playbookRun.owner_user_id}
+                            placeholder={'Assign the owner role'}
+                            placeholderButtonClass={'NoAssignee-button'}
+                            profileButtonClass={'Assigned-button'}
+                            enableEdit={true}
+                            getUsers={fetchUsers}
+                            onSelectedChange={onSelectedProfileChange}
+                            selfIsFirstOption={true}
+                        />
+                    </div>
+                    <div className='first-title'>
+                        {'Duration'}
+                        <Duration
+                            from={props.playbookRun.create_at}
+                            to={props.playbookRun.end_at}
+                        />
+                    </div>
+                </div>
+            </Row>
+            <div id={'playbookRunRHSUpdates'}>
+                <div className='title'>
+                    {'Recent update:'}
+                </div>
+                <PostCard post={latestUpdatePost}/>
             </div>
-        </Scrollbars>
+        </div>
     );
 };
+
+const Description = styled.div`
+    padding: 0 0 14px 0;
+`;
+
+const Row = styled.div`
+    padding: 0 0 24px 0;
+`;
+
+const NoDescription = styled.div`
+    color: rgba(var(--center-channel-color-rgb), 0.64);
+    margin-bottom: 10px;
+`;
 
 export default RHSAbout;
